+++ skygear/src/main/java/io/skygear/skygear/AuthContainer.java
@@ -270,6 +270,152 @@ public class AuthContainer implements AuthResolver {
         this.getContainer().callLambdaFunction("user:reset-password", argv, handler);
     }
 
+    /**
+     * Sets admin roles.
+     *
+     * @param role    the roles array
+     * @param handler the handler
+     */
+    public void setAdminRole(Role[] role, SetRoleResponseHandler handler) {
+        SetAdminRoleRequest request = new SetAdminRoleRequest(role);
+        request.responseHandler = handler;
+
+        RequestManager requestManager = this.getContainer().requestManager;
+        requestManager.sendRequest(request);
+    }
+
+    /**
+     * Sets an admin role.
+     *
+     * @param role    the role
+     * @param handler the handler
+     */
+    public void setAdminRole(Role role, SetRoleResponseHandler handler) {
+        this.setAdminRole(new Role[] { role }, handler);
+    }
+
+    /**
+     * Sets default roles.
+     *
+     * @param role    the role array
+     * @param handler the handler
+     */
+    public void setDefaultRole(Role[] role, SetRoleResponseHandler handler) {
+        SetDefaultRoleRequest request = new SetDefaultRoleRequest(role);
+        request.responseHandler = handler;
+
+        RequestManager requestManager = this.getContainer().requestManager;
+        requestManager.sendRequest(request);
+    }
+
+    /**
+     * Sets a default role.
+     *
+     * @param role    the role
+     * @param handler the handler
+     */
+    public void setDefaultRole(Role role, SetRoleResponseHandler handler) {
+        this.setDefaultRole(new Role[] { role }, handler);
+    }
+
+    /**
+     * Get user role
+     *
+     * @param users   the users array
+     * @param handler the handler
+     */
+    public void fetchUserRole(Record[] users, FetchUserRoleResponseHandler handler) {
+        this.fetchUserRole(this.getUserIDs(users), handler);
+    }
+
+    /**
+     * Get user role
+     *
+     * @param userIDs the user id array
+     * @param handler the handler
+     */
+    public void fetchUserRole(String[] userIDs, FetchUserRoleResponseHandler handler) {
+        FetchUserRoleRequest request = new FetchUserRoleRequest(userIDs);
+        request.responseHandler = handler;
+
+        RequestManager requestManager = this.getContainer().requestManager;
+        requestManager.sendRequest(request);
+    }
+
+    /**
+     * Assign user role
+     *
+     * @param users   the users array
+     * @param roles   the roles array
+     * @param handler the handler
+     */
+    public void assignUserRole(Record[] users, Role[] roles, SetUserRoleResponseHandler handler) {
+        this.assignUserRole(this.getUserIDs(users), this.getRoleNames(roles), handler);
+    }
+
+    /**
+     * Assign user role
+     *
+     * @param userIDs   the user id array
+     * @param roleNames the role name array
+     * @param handler   the handler
+     */
+    public void assignUserRole(String[] userIDs, String[] roleNames, SetUserRoleResponseHandler handler) {
+        AssignUserRoleRequest request = new AssignUserRoleRequest(userIDs, roleNames);
+        request.responseHandler = handler;
+
+        RequestManager requestManager = this.getContainer().requestManager;
+        requestManager.sendRequest(request);
+    }
+
+    /**
+     * Assign user role
+     *
+     * @param users   the users array
+     * @param roles   the roles array
+     * @param handler the handler
+     */
+    public void revokeUserRole(Record[] users, Role[] roles, SetUserRoleResponseHandler handler) {
+        this.revokeUserRole(this.getUserIDs(users), this.getRoleNames(roles), handler);
+    }
+
+    /**
+     * Revoke user role
+     *
+     * @param userIDs   the user id array
+     * @param roleNames the role name array
+     * @param handler   the handler
+     */
+    public void revokeUserRole(String[] userIDs, String[] roleNames, SetUserRoleResponseHandler handler) {
+        RevokeUserRoleRequest request = new RevokeUserRoleRequest(userIDs, roleNames);
+        request.responseHandler = handler;
+
+        RequestManager requestManager = this.getContainer().requestManager;
+        requestManager.sendRequest(request);
+    }
+
+    private String[] getUserIDs(Record[] users) {
+        String[] userIDs = new String[users.length];
+        for (int i = 0; i < users.length; i++) {
+            Record user = users[i];
+            if (!user.getType().equals("user")) {
+                throw new InvalidParameterException("Record type should be user");
+            }
+
+            userIDs[i] = users[i].getId();
+        }
+
+        return userIDs;
+    }
+
+    private String[] getRoleNames(Role[] roles) {
+        String[] roleNames = new String[roles.length];
+        for (int i = 0; i < roles.length; i++) {
+            roleNames[i] = roles[i].getName();
+        }
+
+        return roleNames;
+    }
 
     @Override
     public void resolveAuthUser(Record user, String accessToken) {
